+++ NEWS.md
@@ -1,7 +1,12 @@
+# tvthemes 1.2.1
+
+* Fix tests, re: check environment
+
 # tvthemes 1.2.0
 
 * Add Github Actions
   * R-CMD-Check, codecov, build pkgdown
+* Add license for 'Some Time Later' font
 
 # tvthemes 1.1.0